refactor(Discovery v1): Add back handwritten query classes to avoid breaking changes
