+++ discovery/src/test/java/com/ibm/watson/discovery/v1/DiscoveryServiceIT.java
@@ -131,8 +131,8 @@ import com.ibm.watson.discovery.v1.model.UpdateCredentialsOptions;
 import com.ibm.watson.discovery.v1.model.UpdateDocumentOptions;
 import com.ibm.watson.discovery.v1.model.UpdateEnvironmentOptions;
 import com.ibm.watson.discovery.v1.model.UpdateTrainingExampleOptions;
-import com.ibm.watson.discovery.query.AggregationType;
-import com.ibm.watson.discovery.query.Operator;
+import com.ibm.watson.discovery.v1.query.AggregationType;
+import com.ibm.watson.discovery.v1.query.Operator;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Assume;
