+++ discovery/src/test/java/com/ibm/watson/discovery/v1/DiscoveryServiceTest.java
@@ -124,8 +124,8 @@ import com.ibm.watson.discovery.v1.model.UpdateCredentialsOptions;
 import com.ibm.watson.discovery.v1.model.UpdateDocumentOptions;
 import com.ibm.watson.discovery.v1.model.UpdateEnvironmentOptions;
 import com.ibm.watson.discovery.v1.model.UpdateTrainingExampleOptions;
-import com.ibm.watson.discovery.query.AggregationType;
-import com.ibm.watson.discovery.query.Operator;
+import com.ibm.watson.discovery.v1.query.AggregationType;
+import com.ibm.watson.discovery.v1.query.Operator;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.RecordedRequest;
 import org.junit.After;
